backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
tensorflow==2.20.0
numpy==2.3.1
pillow==11.3.0
scipy==1.16.1
python-multipart==0.0.20
